+++ examples/signature-replay/test/VaultPool.t.sol
@@ -8,23 +8,23 @@ import {VaultPool, IERC20} from "../src/VaultPool.sol";
 contract VaultPoolTest is Test {
     MockERC20 token;
     VaultPool pool;
-    uint256   signerKey = 0xA11CE;
-    address   signer;
-    address   player = makeAddr("player");
-    uint256   AMOUNT;
-    uint256   THRESHOLD;
+    uint256 signerKey = 0xA11CE;
+    address signer;
+    address player = makeAddr("player");
+    uint256 AMOUNT;
+    uint256 THRESHOLD;
 
     function setUp() public {
         signer = vm.addr(signerKey);
         token = new MockERC20("Vault Token", "VLT", address(this), 1_000_000 ether);
-        pool  = new VaultPool(signer, IERC20(address(token)));
+        pool = new VaultPool(signer, IERC20(address(token)));
         token.transfer(address(pool), 1_000_000 ether);
-        AMOUNT    = pool.AUTHORIZED_AMOUNT();
+        AMOUNT = pool.AUTHORIZED_AMOUNT();
         THRESHOLD = pool.SOLVE_THRESHOLD();
     }
 
     function _sign(address to, uint256 amount) internal view returns (bytes memory) {
-        bytes32 inner  = keccak256(abi.encodePacked("Withdraw", to, amount));
+        bytes32 inner = keccak256(abi.encodePacked("Withdraw", to, amount));
         bytes32 digest = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", inner));
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(signerKey, digest);
         return abi.encodePacked(r, s, v);
@@ -40,7 +40,7 @@ contract VaultPoolTest is Test {
     function test_ReplayDrainsToSolve() public {
         bytes memory sig = _sign(player, AMOUNT);
         vm.startPrank(player);
-        for (uint i = 0; i < 10; i++) {
+        for (uint256 i = 0; i < 10; i++) {
             pool.withdraw(player, AMOUNT, sig);
         }
         vm.stopPrank();
